(refactor) 패키지이름 변경, LottoStatistics 피드백 반영
