+++ src/main/java/lotto/lottogame/domain/Lotteries.java
@@ -1,4 +1,4 @@
-package lotto.step3_4.domain;
+package lotto.lottogame.domain;
 
 import java.util.ArrayList;
 import java.util.List;
