+++ src/main/java/lotto/lottogame/domain/OrderPrice.java
@@ -1,4 +1,4 @@
-package lotto.step3_4.domain;
+package lotto.lottogame.domain;
 
 import java.util.Objects;
