+++ src/test/java/lotto/lottogame/domain/LottoGeneratorTest.java
@@ -1,4 +1,4 @@
-package lotto.step3_4.domain;
+package lotto.lottogame.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
